third commit

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,24 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use db;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=>['index','show']]);
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -39,13 +52,34 @@ class PostsController extends Controller
     {
         $this->validate($request,[
             'title'=>'required',
-            'body'=>'required'
+            'body'=>'required',
+            'cover_image'=>'image|max:1999'
         ]);
+
+        // handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get the file with extension
+            $filenameWithExt=$request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename=pathinfo($filenameWithExt,PATHINFO_FILENAME);
+            // Get just ext
+            $extension=$request->file('cover_image')->getClientOriginalExtension();
+            // FIle name to store
+            $fileNameToStore=$filename.'_'.time().'.'.$extension;
+            //Upload Image
+            $path=$request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+        else{
+            $fileNameToStore='noimage.jpg';
+        }
+
+
         //create post
         $post=new Post;
         $post->title=$request->input('title');
         $post->body=$request->input('body');
         $post->user_id=auth()->user()->id;
+        $post->cover_image=$fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success','Post created');
@@ -71,8 +105,15 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+        
         $post= Post::find($id);
+        // check for correct user
+        if(auth()->user()->id!==$post->user_id)
+        {
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         return view('posts.edit')->with('post',$post);
+
     }
 
     /**
@@ -88,10 +129,29 @@ class PostsController extends Controller
             'title'=>'required',
             'body'=>'required'
         ]);
+
+        // handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get the file with extension
+            $filenameWithExt=$request->file('cover_image')->getClientOriginalName();
+            //Get just filename
+            $filename=pathinfo($filenameWithExt,PATHINFO_FILENAME);
+            // Get just ext
+            $extension=$request->file('cover_image')->getClientOriginalExtension();
+            // FIle name to store
+            $fileNameToStore=$filename.'_'.time().'.'.$extension;
+            //Upload Image
+            $path=$request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+       
+
         //create post
         $post=Post::find($id);
         $post->title=$request->input('title');
         $post->body=$request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success','Post Updated');
@@ -106,6 +166,18 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post=POST::find($id);
+
+        // check for correct user
+        if(auth()->user()->id!==$post->user_id)
+        {
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
+        if($post->cover_image!='noimage.jpg'){
+            //Delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+
+        }
+        
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
